feat(styleguide): update components
closes: #408
