+++ src/components/__tests__/Image.test.js
@@ -8,7 +8,7 @@ describe('Image Component', () => {
       <Image
         logo={
           {
-            src: '/1.png',
+            src: 'http://linewalks.com/1.png',
             alt: '',
             width: 237,
             height: 63,
@@ -16,7 +16,7 @@ describe('Image Component', () => {
         }
       />,
     )
-    expect(wrapper.html()).toBe(`<img alt="" src="/1.png" srcset="/1.png 1x, /1@2x.png 2x" width="237px" height="63px">`)
+    expect(wrapper.html()).toBe(`<img alt="" src="http://linewalks.com/1.png" srcset="http://linewalks.com/1.png 1x, http://linewalks.com/1@2x.png 2x" width="237px" height="63px">`)
   })
 
   it('width, height 이 없는 경우', () => {
