+++ src/components/button/Button.test.js
@@ -3,7 +3,7 @@ import { shallow, mount } from 'enzyme';
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
 import _ from 'lodash'
-import Button, { ButtonLink, ButtonTextLink } from '@Components/button/Button'
+import Button from '@Components/button/Button'
 
 const toHaveStyleRules = (component, property, options) => {
   let hyphen = ''
@@ -219,87 +219,6 @@ describe('Button Size', () => {
 
     delete rules.img
     delete rules.marginRight
-
-    toHaveStyleRules(tree, rules)
-  })
-})
-
-it('ButtonLink', () => {
-  const tree = renderer.create(<ButtonLink variant="basic_line" />).toJSON()
-  const rules = {
-    fontWeight: 'bold',
-    fontSize: '14px',
-    minWidth: 'auto',
-    paddingLeft: '8px',
-    paddingRight: '8px',
-    display: 'inline-block',
-    boxSizing: 'border-box',
-    color: '#189bff',
-    hover: {
-      color: '#0070c6',
-    },
-    firstChild: {
-      paddingLeft: '0',
-    },
-  }
-
-  expect(tree).toHaveStyleRule('color', rules.hover.color, {
-    modifier: ':hover',
-  })
-
-  expect(tree).toHaveStyleRule('padding-left', rules.firstChild.paddingLeft, {
-    modifier: ':first-child',
-  })
-
-  delete rules.hover
-  delete rules.firstChild
-
-  toHaveStyleRules(tree, rules)
-})
-
-describe('ButtonLink', () => {
-  it('color & size', () => {
-    const tree = renderer.create(<ButtonLink />).toJSON()
-    const rules = {
-      fontSize: '14px',
-      minWidth: 'auto',
-      paddingLeft: '8px',
-      paddingRight: '8px',
-      display: 'inline-block',
-      boxSizing: 'border-box',
-      color: '#189bff',
-      hover: {
-        color: '#0070c6',
-      },
-      firstChild: {
-        paddingLeft: '0',
-      },
-    }
-
-    expect(tree).toHaveStyleRule('color', rules.hover.color, {
-      modifier: ':hover',
-    })
-
-    expect(tree).toHaveStyleRule('padding-left', rules.firstChild.paddingLeft, {
-      modifier: ':first-child',
-    })
-
-    delete rules.hover
-    delete rules.firstChild
-
-    toHaveStyleRules(tree, rules)
-  })
-})
-
-describe('ButtonTextLinkTag', () => {
-  it('color & size', () => {
-    const tree = renderer.create(<ButtonTextLink />).toJSON()
-    const rules = {
-      fontSize: '16px',
-      fontWeight: 'bold',
-      color: 'rgba(0,0,0,0.6)',
-      textDecoration: 'underline',
-    }
 
     toHaveStyleRules(tree, rules)
   })
