+++ src/components/button/ButtonTextLink.md
@@ -1,0 +1,5 @@
+ButtonTextLink example:
+
+```js
+<ButtonTextLink pizza="🍕">11</ButtonTextLink>
+```
