+++ src/components/button/ToggleButton.md
@@ -1,0 +1,16 @@
+ToggleButton example:
+
+```js
+<ToggleButton
+  data={[
+    {
+      type: 'AA',
+      text: '환자군',
+    },
+    {
+      type: 'BB',
+      text: '진단수',
+    }
+  ]}
+/>
+```
