+++ src/components/charts/RadiusGauge.md
@@ -1,0 +1,5 @@
+RadiusGauge example:
+
+```js
+<RadiusGauge width={300} height={200} score={0} />
+```
