+++ src/components/common/cdmCommon.md
@@ -1,0 +1,16 @@
+LegendList example:
+
+```js
+<LegendList
+  data={[
+    {
+      color: 'red',
+      text: '환자군',
+    },
+    {
+      color: 'blue',
+      text: '진단수',
+    }
+  ]}
+/>
+```
