+++ src/components/form/SelectBox.md
@@ -1,0 +1,10 @@
+SelectBox example:
+
+```js
+<SelectBox>
+  <select>
+    <option value="concept_id">Condition concepts</option>
+    <option value="kcd">KCD</option>
+  </select>
+</SelectBox>
+```
