+++ src/components/layout/Footer.md
@@ -1,0 +1,7 @@
+Footer example:
+
+```js
+<div style={{height: 100, position: 'relative'}}>
+  <Footer />
+</div>
+```
