+++ src/components/layout/Heading.md
@@ -1,0 +1,8 @@
+Heading example:
+
+```js
+<div>
+  <Heading>MDwalks-UI Heading</Heading>
+  <Heading size="15" style={{color: '#0B5FFF'}}>MDwalks-UI Heading</Heading>
+</div>
+```
