+++ src/components/layout/Image.md
@@ -1,0 +1,12 @@
+Image example:
+
+```js
+<Image
+  logo={
+    {
+      src: 'http://retinaimag.es/img/settings-small.png',
+      alt: '',
+    }
+  }
+/>
+```
