+++ src/components/layout/Navbar.md
@@ -1,0 +1,7 @@
+Navbar example:
+
+```js
+<Navbar>
+  Navbar
+</Navbar>
+```
