+++ src/components/pagination/Pagination.md
@@ -1,0 +1,9 @@
+Pagination example:
+
+```js
+const PaginationExample = () => (
+  <Pagination selectPage={1} totalPage={6} drawPageCnt={5} />
+)
+
+PaginationExample()
+```
