+++ src/components/paper/Heatmap.md
@@ -1,0 +1,9 @@
+Heatmap(Beta) example:
+
+```js
+const HeatmapExample = () => (
+  <Heatmap />
+)
+
+HeatmapExample()
+```
