+++ src/components/paper/Sequence.md
@@ -1,0 +1,9 @@
+Sequence(Beta) example:
+
+```js
+const SequenceExample = () => (
+  <Sequence />
+)
+
+SequenceExample()
+```
