+++ src/components/table/Descriptions.md
@@ -1,0 +1,18 @@
+Descriptions example:
+
+```js
+const DescriptionsExample = () => (
+  <Descriptions
+    cellCount={2}
+    data={[
+      {'Drug Count':	'1231'},
+      {'Unique Drug Count':	'123'},
+      {'Condition Count':	'999999'},
+      {'Unique Condition Count': '8888'},
+      {'Patient Count': '7777'}
+    ]}
+  />
+)
+
+DescriptionsExample()
+```
