+++ src/components/table/EmptyPlaceHolder.md
@@ -1,0 +1,9 @@
+EmptyPlaceHolder example:
+
+```js
+const EmptyPlaceHolderExample = () => (
+  <EmptyPlaceHolder />
+)
+
+EmptyPlaceHolderExample()
+```
