+++ src/docz/Button.mdx
@@ -4,7 +4,9 @@ edit: false
 ---
 
 import { Playground, Props } from 'docz'
-import Button, {ButtonLink, ButtonTextLink} from '@Components/button/Button'
+import Button from '@Components/button/Button'
+import ButtonLink from '@Components/button/ButtonLink'
+import ButtonTextLink from '@Components/button/ButtonTextLink'
 
 ## default
 <Playground>
