+++ src/docz/Layout.mdx
@@ -38,7 +38,6 @@ support property style, size
   <Heading size="15" style={{color: '#0B5FFF'}}>MDwalks-UI Heading</Heading>
 </Playground>
 
-
 ## Navbar
 
 support property style
