+++ src/index.js
@@ -22,7 +22,9 @@ import RadioList from '@Components/list/RadioList'
 import RadarChart from '@Components/charts/RadarChart'
 import TreeMap from '@Components/charts/TreeMap'
 import TimeToEvent from '@Components/charts/TimeToEvent'
-import Button, { ButtonLink, ButtonTextLink } from '@Components/button/Button'
+import Button from '@Components/button/Button'
+import ButtonLink from '@Components/button/ButtonLink'
+import ButtonTextLink from '@Components/button/ButtonTextLink'
 import ToastCtr from '@Components/toast/ToastCtr'
 import Pagination from '@Components/pagination/Pagination'
 import ToggleButton from '@Components/button/ToggleButton'
