+++ styleguide.config.js
@@ -1,6 +1,8 @@
 const path = require('path')
 const { version } = require('./package');
 
+const ignore = ['**/__tests__/**', '**/*.test.{js,jsx,ts,tsx}', '**/*.spec.{js,jsx,ts,tsx}', '**/*.d.ts']
+
 module.exports = {
   require: [path.resolve(__dirname, 'src/assets/styles/reset.css')],
   moduleAliases: {
@@ -10,4 +12,13 @@ module.exports = {
     '@Cards': path.resolve(__dirname, 'src/components/card'),
   },
   version,
+  ignore: ignore.concat([
+    'src/components/table/THead.js',
+    'src/components/table/TBody.js',
+    'src/components/table/TFoot.js',
+    'src/components/toast/ToastList.js',
+    'src/components/list/Item.js',
+    'src/components/button/utility.js',
+    'src/components/toast/ToastCtr.js',
+  ]),
 }
